Completely rewrite TLV handling
I wasn't happy with the previous TLV handling interface, especially how
it worked in match statements. This is a complete rewrite to one class
per TLV type and decoding of the TLV values in the decoding step instead
of always returning bytes.

Hit a bug in mypy with this though, so this also includes a small mypy
plugin to work around python/mypy#12886