Implement pairing flow
This implements the rest of the pairing flow, although without any
persistence for the private key or pairing information.

This also removes the ASGI layer. HomeKit requires persistent, stateful
connections which turned out to be a poor fit for the ASGI abstraction
layer and the server wouldn't be useable with a standard ASGI server
anyway.